Linear interpolation

+++ src/gaussinterpolator.h
@@ -9,7 +9,7 @@
 
 class GaussianInterpolator : public Interpolator {
 public:
-	HdrImage calculateInterpolationOn(double observation);
+	HdrImage calculateInterpolationOn(int observation);
 	HdrImage takeStep();
 };
